feat: 시간 위젯 수정 및 페이지 완성
