+++ lib/core/router.dart
@@ -1,101 +1,133 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
-import 'package:mongbi_app/core/custom_transition_page.dart';
-import 'package:mongbi_app/core/main_scaffold.dart';
-import 'package:mongbi_app/presentation/auth/social_login_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_analysis_loading_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_analysis_result_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_interpretation_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_write_page.dart';
-import 'package:mongbi_app/presentation/history/history_page.dart';
-import 'package:mongbi_app/presentation/home/home_page.dart';
-import 'package:mongbi_app/presentation/remind/remind_time_setting_page.dart';
-import 'package:mongbi_app/presentation/remind/remind_time_time_setting_page.dart';
-import 'package:mongbi_app/presentation/setting/alarm_setting_page.dart';
-import 'package:mongbi_app/presentation/setting/nickname_input_page.dart';
-import 'package:mongbi_app/presentation/setting/profile_setting_page.dart';
-import 'package:mongbi_app/presentation/setting/setting_page.dart';
-import 'package:mongbi_app/presentation/splash/splash_page.dart';
-import 'package:mongbi_app/presentation/statistics/statistics_page.dart';
+import 'package:mongbi_app/core/font.dart';
+import 'package:mongbi_app/presentation/remind/view_model/remind_time_setting_view_model.dart';
+import 'package:mongbi_app/presentation/remind/widgets/remind_time_setting_button_widget.dart';
+import 'package:mongbi_app/presentation/remind/widgets/remind_time_setting_text_widget.dart';
+import 'package:mongbi_app/presentation/remind/widgets/remind_time_setting_widget.dart';
+import 'package:permission_handler/permission_handler.dart';
 
-final GoRouter router = GoRouter(
-  initialLocation: '/',
-  routes: [
-    GoRoute(path: '/', redirect: (_, __) => '/remindtime_setting'),
+class RemindTimePickerPage extends StatefulWidget {
+  const RemindTimePickerPage({super.key});
 
-    ShellRoute(
-      builder: (context, state, child) {
-        return MainScaffold(child: child);
-      },
-      routes: [
-        GoRoute(path: '/home', builder: (context, state) => const HomePage()),
-        GoRoute(
-          path: '/history',
-          builder: (context, state) => const HistoryPage(),
+  @override
+  State<RemindTimePickerPage> createState() => _RemindTimePickerPageState();
+}
+
+class _RemindTimePickerPageState extends State<RemindTimePickerPage> {
+  TimeOfDay selectedTime = const TimeOfDay(hour: 8, minute: 0);
+
+  @override
+  void initState() {
+    super.initState();
+    // 앱 최초 진입 시 알림 권한 요청
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      final granted =
+          await NotificationService().requestNotificationPermission();
+      if (!granted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(content: Text('알림 권한이 거부되었습니다. 설정에서 허용해주세요.')),
+        );
+      }
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color(0xFFFAFAFA),
+      appBar: AppBar(
+        backgroundColor: const Color(0xFFFAFAFA),
+        surfaceTintColor: Colors.transparent,
+        elevation: 0,
+        leading: Padding(
+          padding: const EdgeInsets.only(left: 24, top: 16),
+          child: IconButton(
+            icon: SvgPicture.asset('assets/icons/back-arrow.svg'),
+            onPressed: () => Navigator.pop(context),
+            splashRadius: 20,
+          ),
         ),
-        GoRoute(
-          path: '/statistics',
-          builder: (context, state) => const StatisticsPage(),
+        foregroundColor: Colors.black,
+        toolbarHeight: 56,
+        titleSpacing: 0,
+      ),
+      body: SafeArea(
+        child: Stack(
+          children: [
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 24),
+              child: Column(
+                children: [
+                  const SizedBox(height: 72),
+                  const OnboardingText(
+                    text: '몽비와 함께할 시간을 알려줘!',
+                    type: RemindTimeSettingTextType.title,
+                  ),
+                  const SizedBox(height: 8),
+                  const OnboardingText(
+                    text: '원하는 리마인드 시간을\n설정해주세요.',
+                    type: RemindTimeSettingTextType.description,
+                    align: TextAlign.center,
+                  ),
+                  const SizedBox(height: 89),
+                  SizedBox(
+                    height: 188,
+                    child: CustomTimePicker(
+                      onChanged: (time, amPm) {
+                        selectedTime = time;
+                      },
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Positioned(
+              bottom: 108,
+              left: 0,
+              right: 0,
+              child: Center(
+                child: Text(
+                  '나중에 다시 설정할 수 있어요',
+                  style: Font.subTitle14.copyWith(
+                    color: const Color(0xFFA6A1AA),
+                  ),
+                ),
+              ),
+            ),
+            Positioned(
+              bottom: 32,
+              left: 24,
+              right: 24,
+              child: RemindTimeSettingButtonWidget(
+                onTap: () async {
+                  try {
+                    await NotificationService().scheduleDailyReminder(
+                      selectedTime,
+                    );
+                    context.go('/home');
+                  } catch (e) {
+                    if (e is PlatformException &&
+                        e.code == 'exact_alarms_not_permitted') {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        const SnackBar(
+                          content: Text('정확한 알람 권한이 필요해요! 설정에서 허용해주세요.'),
+                        ),
+                      );
+                      await NotificationService()
+                          .openExactAlarmSettingsIfNeeded();
+                    } else {
+                     
+                    }
+                  }
+                },
+              ),
+            ),
+          ],
         ),
-        GoRoute(path: '/setting', builder: (context, state) => SettingPage()),
-      ],
-    ),
-    GoRoute(path: '/splash', builder: (context, state) => SplashPage()),
-    GoRoute(
-      path: '/social_login',
-      builder: (context, state) => SocialLoginPage(),
-    ),
-    GoRoute(
-      path: '/nickname_input',
-      builder: (context, state) => NicknameInputPage(),
-    ),
-    GoRoute(
-      path: '/remindtime_setting',
-      builder: (context, state) => RemindTimeSettingPage(),
-    ),
-        GoRoute(
-      path: '/remindtime_setting',
-      builder: (context, state) => RemindTimeSettingPage(),
-    ),
-    GoRoute(
-      path: '/remindtime_time_setting',
-      builder: (context, state) => RemindTimePickerPage(),
-    ),
-    
-    GoRoute(
-      path: '/dream_write',
-      pageBuilder:
-          (context, state) => buildFadeTransitionPage(
-            key: state.pageKey,
-            child: DreamWritePage(),
-          ),
-    ),
-    GoRoute(
-      path: '/dream_analysis_loading',
-      pageBuilder:
-          (context, state) => buildFadeTransitionPage(
-            key: state.pageKey,
-            child: DreamAnalysisLoadingPage(),
-          ),
-    ),
-    GoRoute(
-      path: '/dream_analysis_result',
-      pageBuilder:
-          (context, state) => buildFadeTransitionPage(
-            key: state.pageKey,
-            child: DreamAnalysisResultPage(),
-          ),
-    ),
-    GoRoute(
-      path: '/dream_interpretation',
-      builder: (context, state) => DreamInterpretationPage(),
-    ),
-    GoRoute(
-      path: '/profile_setting',
-      builder: (context, state) => ProfileSettingPage(),
-    ),
-    GoRoute(
-      path: '/alarm_setting',
-      builder: (context, state) => AlarmSettingPage(),
-    ),
-  ],
-);
+      ),
+    );
+  }
+}
